% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectRunFromFile.R
\name{runProjectionExcel}
\alias{runProjectionExcel}
\title{Automate simulations using an excel input file}
\usage{
runProjectionExcel(inputDir, inputFile, outputDir)
}
\arguments{
\item{inputDir}{Where to look for the input file.}

\item{inputFile}{The excel input file}

\item{outputDir}{Where to put the output files}
}
\description{
FishSimGTG wrapper to automate multiple simulations using an excel input file
}
