% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimTools.R
\docType{package}
\name{fishSimTools}
\alias{fishSimTools}
\alias{fishSimTools-package}
\title{Toolkit for using fishSimGTG}
\description{
Toolkit for effecient analysis and sharing of results produced by fishSimGTG.
}
\author{
\strong{Maintainer}: William Harford \email{bill@natureanalytics.ca}

}
